refactor(ProductList): ProductModel에서 count제거, CartModel에 count 추가

+++ app/src/main/java/woowacourse/shopping/data/respository/cart/CartRepositoryImpl.kt
@@ -13,7 +13,7 @@ class CartRepositoryImpl(
     override fun addCartProduct(
         productId: Long,
         onFailure: () -> Unit,
-        onSuccess: () -> Unit,
+        onSuccess: (Long) -> Unit,
     ) {
         cartRemoteDataSource.requestPostCartItem(productId, onFailure, onSuccess)
     }
